refactor: move map stats to geoUtils; export all data without config filter-out
Move mapStatisticsFromFeatureCollection into geoUtils. For export, do not
apply config-based filter-out so users get all data when they download;
add comment explaining the requirement.
